(chore) Consolidate parsing function into parse.c file
