+++ src/parse/parse.c
@@ -6,7 +6,7 @@
 /*   By: julberna <julberna@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/12 17:15:02 by julberna          #+#    #+#             */
-/*   Updated: 2024/04/12 21:51:01 by julberna         ###   ########.fr       */
+/*   Updated: 2024/04/13 16:18:52 by julberna         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,4 +28,48 @@ void	check_input(t_game *cub, int argc, char **argv)
 		cuberror("Wrong number of arguments.", cub);
 	cub->map_path = argv[1];
 	cub->map_data.raw_data = get_raw_data(cub->map_path, cub);
+}
+
+char	**get_raw_data(char *map_path, t_game *cub)
+{
+	int		i;
+	int		fd;
+	int		rows;
+	char	*temp_line;
+	char	**raw_data;
+
+	i = 0;
+	fd = open(map_path, O_RDONLY);
+	if (fd == -1)
+		cuberror("Could not open map file.", cub);
+	rows = count_rows(fd);
+	close(fd);
+	fd = open(map_path, O_RDONLY);
+	raw_data = (char **)ft_calloc(rows + 1, sizeof(char *));
+	while (i < rows)
+	{
+		temp_line = get_next_line(fd);
+		raw_data[i] = (char *)ft_calloc(ft_strlen(temp_line) + 1, sizeof(char));
+		if (raw_data[i] && temp_line)
+			ft_strlcpy(raw_data[i], temp_line, ft_strlen(temp_line) + 1);
+		i++;
+		free(temp_line);
+	}
+	return (raw_data);
+}
+
+int	count_rows(int fd)
+{
+	int		rows;
+	char	*temp_line;
+
+	rows = 0;
+	temp_line = get_next_line(fd);
+	while (temp_line)
+	{
+		rows++;
+		free(temp_line);
+		temp_line = get_next_line(fd);
+	}
+	return (rows);
 }
